simplify new a little bit
